Rename: atoms/Chat 안에 있던 UserChat과 MyChat을 atoms 폴더로 경로 변경
